+++ notification/doom9.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python3
+import requests
+import sys
+import http.cookiejar
+
+session = requests.Session()
+
+# doom9 login is hard